# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CertificateData": {
        "description": "Certificate data (service-layer response model).",
        "properties": {
          "certificate_type": {
            "title": "Certificate Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "phases_completed": {
            "title": "Phases Completed",
            "type": "integer"
          },
          "recipient_name": {
            "title": "Recipient Name",
            "type": "string"
          },
          "total_phases": {
            "title": "Total Phases",
            "type": "integer"
          },
          "verification_code": {
            "title": "Verification Code",
            "type": "string"
          }
        },
        "required": [
          "id",
          "certificate_type",
          "verification_code",
          "recipient_name",
          "issued_at",
          "phases_completed",
          "total_phases"
        ],
        "title": "CertificateData",
        "type": "object"
      },
...
      "CertificateRequest": {
        "description": "Request to generate a certificate.",
        "properties": {
          "certificate_type": {
            "default": "full_completion",
            "pattern": "^full_completion$",
            "title": "Certificate Type",
            "type": "string"
          },
          "recipient_name": {
            "maxLength": 100,
            "minLength": 2,
            "title": "Recipient Name",
            "type": "string"
          }
        },
        "required": [
          "recipient_name"
        ],
        "title": "CertificateRequest",
        "type": "object"
      },
      "CertificateVerifyResponse": {
        "description": "Response for certificate verification.",
        "properties": {
          "certificate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CertificateData"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_valid": {
            "title": "Is Valid",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "is_valid",
          "message"
        ],
        "title": "CertificateVerifyResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UserCertificatesResponse": {
        "description": "All certificates for a user.",
        "properties": {
          "certificates": {
            "items": {
              "$ref": "#/components/schemas/CertificateData"
            },
            "title": "Certificates",
            "type": "array"
          },
          "full_completion_eligible": {
            "title": "Full Completion Eligible",
            "type": "boolean"
          }
        },
        "required": [
          "certificates",
          "full_completion_eligible"
        ],
        "title": "UserCertificatesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Learn to Cloud API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/certificates": {
      "get": {
        "description": "Get all certificates for the authenticated user.",
        "operationId": "get_user_certificates_api_certificates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCertificatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Get User Certificates",
        "tags": [
          "certificates"
        ]
      },
      "post": {
        "description": "Generate a completion certificate for eligible users.",
        "operationId": "generate_certificate_endpoint_api_certificates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateData"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "User not eligible for certificate"
          },
          "409": {
            "description": "Certificate already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Certificate Endpoint",
        "tags": [
          "certificates"
        ]
      }
    },
...
    "/api/certificates/verify/{verification_code}": {
      "get": {
        "description": "Verify a certificate by its verification code (public endpoint).",
        "operationId": "verify_certificate_endpoint_api_certificates_verify__verification_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "verification_code",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 10,
              "title": "Verification Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation error - code too short or long"
          }
        },
        "summary": "Verify Certificate Endpoint",
        "tags": [
          "certificates"
        ]
      }
    },
...
  }
}